requirements.txt
discord.py
mistralai
python-dotenv